por si las moscas

+++ src/Aplicacion/Vistas/Pedido.java
@@ -1348,7 +1348,8 @@ public class Pedido extends View implements IView{
            tipoInsercion="comun";
        }
        //calculando puntos ganados en la compra
-       String puntosCompra = ((Integer.parseInt(jTextField17.getText()))-(Integer.parseInt(jTextField31.getText())))+"";
+       
+       String puntosCompra = ((Float.parseFloat(jTextField17.getText()))-(Float.parseFloat(jTextField31.getText())))+"";
       //insertando en la pedido
         db= new DataBase();
         db.insertar3("INSERT INTO pedido(fecha,horaRegistro,numero,direccion,nota,precio_total,estado,id_usuario,id_empleado,id_cliente,tipo,id_municipio,condonado,puntos,obsequio,puntos_descontados_por_obsequio, bodega) VALUES('"+fechaActual+"',(SELECT CURRENT_TIME()),'"+jTextField8.getText()+"','"+jTextField6.getText()+"','"+jTextArea2.getText()+"','"+jTextField22.getText()+"','"+jComboBox3.getSelectedItem()+"','"+this.usuario+"','"+cedulaEmpleado+"','"+telefono+"','"+tipoInsercion+"','"+jTextField12.getText()+"','"+jTextField19.getText()+"','"+puntosCompra+"','ninguno','0.0','"+bodega+"')");
